+++ CHANGELOG.rst
@@ -1,4 +1,7 @@
 Changelog
 ---------
+Next Release
+  * Look up SRV names in DNS additional records segment
+
 2014-03-11: v0.1.0
   * Initial version